Only scroll to success message if it isn't already visible in the viewport
If the user already has the form in the middle of their browser, it's disorienting for the user when it just auto jumps so the success message as at the top of the viewport. This code checks to see if the success message is visible without scrolling, and doesn't scroll if so.